+++ django/applications/catmaid/static/js/widgets/import-export-widget.js
@@ -89,10 +89,10 @@ annotations, neuron name, connectors or partner neurons.
 </p>
 
 <p>
-<label>SWC file: <input type="file" accept=".swc" data-role="swc-import-file" />
+<label>SWC file: <input type="file" accept=".swc" multiple data-role="swc-import-file" />
 </p>
 <p>
-<button data-role="import-swc">Import SWC</button>
+<button data-role="import-swc">Import SWC file(s)</button>
 </p>
  `;
 
@@ -142,16 +142,54 @@ annotations, neuron name, connectors or partner neurons.
           var $importContainer = $(importContainer);
           $importContainer.find('button[data-role=import-swc]').click(function() {
             var fileInput = $importContainer.find('input[data-role=swc-import-file]');
-            if (fileInput.length !== 1) {
+            if (fileInput.length === 0) {
               CATMAID.warn("No SWC file input found");
               return;
             }
-            fileInput = fileInput[0];
-            if (fileInput.files.length !== 1) {
-              CATMAID.warn("Please select exactly one file");
-              return;
+            var files = fileInput[0].files;
+            var importedFiles = new Map();
+            var failedImports = new Map();
+            var importQueue = Promise.resolve();
+            for (let i=0; i<files.length; ++i) {
+              let file = files[i];
+              importQueue = importQueue
+                .then(function() {
+                  return import_swc(file)
+                    .then(function(data) {
+                      CATMAID.msg("SWC successfully imported", "Neuron ID:" +
+                          data.neuron_id + " Skeleton ID: " + data.skeleton_id);
+                      importedFiles.set(file.name, data);
+                      if (files.length === 1) {
+                        CATMAID.TracingTool.goToNearestInNeuronOrSkeleton(
+                          'skeleton', data.skeleton_id);
+                      }
+                    })
+                    .catch(function(error) {
+                      CATMAID.warn("SWC not imported: " + file.name);
+                      failedImports.set(file.name, error);
+                    });
+                });
             }
-            import_swc(fileInput.files[0]);
+
+            importQueue
+              .then(function() {
+                if (failedImports.size === 0) {
+                  CATMAID.msg("Success", "Imported " + importedFiles.size + " neurons");
+                } else {
+                  var msg;
+                  if (importedFiles === 0) {
+                    msg = "Could not import any selected SWC file";
+                  } else {
+                    msg = "Some SWC files could not be imported";
+                  }
+                  var details = [];
+                  for (var [key, value] of failedImports.entries()) {
+                    details.push("File: " + key + " Error: " + value.error);
+                  }
+                  CATMAID.error(msg, details.join("\n"));
+                }
+              })
+              .catch(CATMAID.handleError);
           });
         }
 
@@ -226,28 +264,27 @@ annotations, neuron name, connectors or partner neurons.
     dialog.show();
   }
 
-  function import_swc(file) {
+  function import_swc(file, autoSelect) {
     if (!file) {
-      CATMAID.warn("Need file");
-      return;
+      return Promise.reject("Need file");
     }
 
     var data = new FormData();
     data.append('file', file);
-    $.ajax({
-        url : CATMAID.makeURL(project.id + "/skeletons/import"),
-        processData : false,
-        contentType : false,
-        type : 'POST',
-        data : data,
-    }).done(function(data) {
-        if (data.skeleton_id) {
-          CATMAID.msg("SWC successfully imported", "Neuron ID:" + data.neuron_id
-            + " Skeleton ID: " + data.skeleton_id);
-          CATMAID.TracingTool.goToNearestInNeuronOrSkeleton('skeleton', data.skeleton_id);
-        } else {
-          CATMAID.error("Something went wrong during the import", data);
-        }
+    return new Promise(function(resolve, reject) {
+      $.ajax({
+          url : CATMAID.makeURL(project.id + "/skeletons/import"),
+          processData : false,
+          contentType : false,
+          type : 'POST',
+          data : data,
+      }).done(function(data) {
+          if (data.skeleton_id) {
+            resolve(data);
+          } else {
+            reject(data);
+          }
+      });
     });
   }